# api/openapi.yaml
openapi: 3.0.3
info:
  title: Synaptic API
  version: 0.0.1
  description: |
    Behavioral contract captured from the NestJS implementation on `main`
    (source of truth: the 7 controllers, guards, interceptors, DTOs and
    `src/main.ts` / `src/app.module.ts` under `src/`). 33 routes.

    ## Request pipeline order (per request)

    1. CORS (credentials enabled, single origin from `CLIENT_URL`,
       default `http://localhost:4200`).
    2. Body parsing (JSON and urlencoded, 5 MB limit; oversized -> 413).
    3. `ThrottlerGuard` (global) -> 429 when the per-IP limit is exceeded.
    4. `CsrfGuard` (global) -> 403 for unsafe methods without a valid
       double-submit token pair.
    5. Route guards: `JwtAuthGuard` -> 401, then `RolesGuard` -> 403.
    6. Global `ValidationPipe` (`whitelist: true`,
       `forbidNonWhitelisted: true`, `transform: true`) -> 400 on schema
       violations, on unknown properties
       (`property <name> should not exist`), and on parameter pipe
       failures (`MongoIdPipe`, `ParseArrayPipe`).
    7. Handler/service errors (404, 409, 503, unhandled -> 500).

    ## Authentication

    JWT access tokens are accepted from EITHER the `Authorization: Bearer`
    header OR the `access_token` cookie (header takes precedence). Tokens
    are signed with `JWT_SECRET` and must match `JWT_ISSUER` /
    `JWT_AUDIENCE`; payload is `{ email, username, sub }`. `sub` must
    reference an existing user or the request is 401. Refresh uses the
...

...

...

...

...
servers:
...
paths:
...
  /categories/category/create:
    post:
      tags: [categories]
      operationId: createCategory
      summary: Create a category
      description: >-
        Admin only. Note the doubled path segments (controller prefix
        `categories` + route `category/create`). A duplicate `slug`
        violates the unique Mongo index and surfaces as 500, not 409.
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/Throttled' }
        '500': { $ref: '#/components/responses/InternalError' }
  /categories/categories:
    get:
      tags: [categories]
      operationId: getCategories
      summary: List all categories
      description: >-
        Any authenticated role. Note the doubled path segments. Sorted by
        `title` ascending.
      security:
        - bearerAuth: []
        - accessCookie: []
      responses:
        '200':
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/Throttled' }
  /categories/{id}:
    get:
      tags: [categories]
      operationId: getCategoryById
      summary: Get a category by ID
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - { $ref: '#/components/parameters/MongoIdParam' }
      responses:
        '200':
          description: The requested category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/InvalidMongoId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Category not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Category not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
    delete:
      tags: [categories]
      operationId: deleteCategory
      summary: Delete a category
      description: Admin only. Returns 204 with an empty body.
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        - { $ref: '#/components/parameters/MongoIdParam' }
      responses:
        '204': { description: Category deleted; empty body }
        '400': { $ref: '#/components/responses/InvalidMongoId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Category not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Category not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
  /topics/create:
    post:
      tags: [topics]
      operationId: createTopic
      summary: Create a topic
      description: >-
        Admin only. `category` must reference an existing category ID
        (not verified at validation time; a dangling ID is persisted and
        only breaks later on populate). A duplicate `slug` surfaces as
        500, not 409.
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTopicRequest' }
      responses:
        '201':
          description: Created topic (category populated)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Topic' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/Throttled' }
        '500': { $ref: '#/components/responses/InternalError' }
  /topics:
    get:
      tags: [topics]
      operationId: getTopics
      summary: List all topics
      description: Any authenticated role. Sorted by `title` ascending.
      security:
        - bearerAuth: []
        - accessCookie: []
      responses:
        '200':
          description: All topics
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Topic' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/Throttled' }
  /topics/{id}:
    get:
      tags: [topics]
      operationId: getTopicById
      summary: Get a topic by ID
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - { $ref: '#/components/parameters/MongoIdParam' }
      responses:
        '200':
          description: The requested topic (category populated)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Topic' }
        '400': { $ref: '#/components/responses/InvalidMongoId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Topic not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Topic not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
    delete:
      tags: [topics]
      operationId: deleteTopic
      summary: Delete a topic
      description: Admin only. Returns 204 with an empty body.
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        - { $ref: '#/components/parameters/MongoIdParam' }
      responses:
        '204': { description: Topic deleted; empty body }
        '400': { $ref: '#/components/responses/InvalidMongoId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Topic not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Topic not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
  /questions/create:
    post:
      tags: [questions]
      operationId: createQuestionSets
      summary: Create question sets (bulk)
      description: >-
        Admin only. Body is a JSON array validated item-by-item via
        `ParseArrayPipe` (a non-array body fails 400; an empty array
        passes and returns `[]`).
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/CreateQuestionSetRequest' }
      responses:
        '201':
          description: Created question sets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/QuestionSet' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/Throttled' }
  /questions/update:
    patch:
      tags: [questions]
      operationId: bulkUpdateQuestionSets
      summary: Update question sets (bulk)
      description: >-
        Admin only. Each item must carry the target `id`; items are
        applied independently via `Promise.all`, so one missing set fails
        the whole request with 404 and sibling updates may already be
        persisted.
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BulkUpdateQuestionSetRequest'
      responses:
        '200':
          description: Updated question sets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/QuestionSet' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Question set not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Question set not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
  /questions/topic/{slug}:
    get:
      tags: [questions]
      operationId: getQuestionSetsByTopicSlug
      summary: List question sets for a topic slug
      description: >-
        Any authenticated role. `populateTopic` is a raw string compare:
        the topic is embedded only when the value is exactly `true`
        (default: no).
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: populateTopic
          in: query
          schema: { type: string, enum: ['true', 'false'], default: 'false' }
      responses:
        '200':
          description: Question sets for the topic (possibly empty)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/QuestionSet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Topic not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Topic not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
  /questions/{id}:
    get:
      tags: [questions]
      operationId: getQuestionSetById
      summary: Get a question set by ID
      description: >-
        Any authenticated role. Note the inverted `populateTopic`
        default compared to the slug route: the topic is embedded unless
        the value is exactly `false`.
      security:
        - bearerAuth: []
        - accessCookie: []
      parameters:
        - { $ref: '#/components/parameters/MongoIdParam' }
        - name: populateTopic
          in: query
          schema: { type: string, enum: ['true', 'false'], default: 'true' }
      responses:
        '200':
          description: The requested question set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuestionSet' }
        '400': { $ref: '#/components/responses/InvalidMongoId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: Question set not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Question set not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
    patch:
      tags: [questions]
      operationId: updateQuestionSet
      summary: Update a question set
      description: Admin only. All body fields optional.
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        - { $ref: '#/components/parameters/MongoIdParam' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateQuestionSetRequest' }
      responses:
        '200':
          description: Updated question set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuestionSet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Question set not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Question set not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
    delete:
      tags: [questions]
      operationId: deleteQuestionSet
      summary: Delete a question set
      description: Admin only. Returns 204 with an empty body.
      security:
        - bearerAuth: []
          csrfCookie: []
          csrfHeader: []
        - accessCookie: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        - { $ref: '#/components/parameters/MongoIdParam' }
      responses:
        '204': { description: Question set deleted; empty body }
        '400': { $ref: '#/components/responses/InvalidMongoId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Question set not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                statusCode: 404
                message: Question set not found
                error: Not Found
        '429': { $ref: '#/components/responses/Throttled' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Checked before the `access_token` cookie.
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly JWT cookie, path=/ (see `x-auth-cookies`).
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly refresh cookie, path=/auth.
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
      description: Must match the `x-csrf-token` header on unsafe methods.
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Must match the `csrf_token` cookie on unsafe methods.
  parameters:
    MongoIdParam:
      name: id
      in: path
      required: true
      description: >-
        MongoDB ObjectId; non-ObjectId values fail `MongoIdPipe` with 400
        `Invalid MongoDB ObjectId`.
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: >-
        DTO validation failed (constraint messages, or `property <name>
        should not exist` for unknown properties)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            statusCode: 400
            message:
              - username must be longer than or equal to 3 characters
              - property role should not exist
            error: Bad Request
    InvalidMongoId:
      description: Path parameter is not a valid MongoDB ObjectId
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            statusCode: 400
            message: Invalid MongoDB ObjectId
            error: Bad Request
    Unauthorized:
      description: Missing, invalid, or expired credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            statusCode: 401
            message: Unauthorized
...
    Forbidden:
      description: CSRF failure or insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            csrf:
              value:
                statusCode: 403
                message: Invalid CSRF token
                error: Forbidden
            role:
              value:
                statusCode: 403
                message: Forbidden resource
                error: Forbidden
    Throttled:
      description: Rate limit exceeded (see `x-throttling`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            statusCode: 429
            message: 'ThrottlerException: Too Many Requests'
...
    InternalError:
      description: >-
        Unhandled failure (e.g. Mongo duplicate-key on a unique `slug`,
        AI provider SDK errors)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            statusCode: 500
            message: Internal server error
  schemas:
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message:
          description: String, or an array of strings for DTO validation.
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        error:
          type: string
          description: >-
            HTTP reason phrase; absent for `UnauthorizedException` and
            other exceptions thrown without an explicit message.
    ObjectId:
      type: string
      pattern: '^[0-9a-fA-F]{24}$'
      example: 665f1e2b9d1a2c3b4d5e6f70
...
    Category:
      type: object
      required: [id, title, slug, description, icon]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        icon: { type: string }
    CreateCategoryRequest:
      type: object
      required: [title, slug, description, icon]
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        icon: { type: string, minLength: 1 }
    Topic:
      type: object
      required: [id, title, slug, description, icon, tags, category]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        icon: { type: string }
        tags:
          type: array
          minItems: 1
          maxItems: 2
          items: { type: string }
        category: { $ref: '#/components/schemas/Category' }
    CreateTopicRequest:
      type: object
      required: [title, slug, description, icon, tags, category]
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        icon: { type: string, minLength: 1 }
        tags:
          type: array
          minItems: 1
          maxItems: 2
          items: { type: string, minLength: 1 }
        category: { $ref: '#/components/schemas/ObjectId' }
    QuestionType:
      type: string
      enum: [mcq, written]
    QuestionOption:
      type: object
      required: [id, text]
      properties:
        id: { type: string, minLength: 1 }
        text: { type: string, minLength: 1 }
    QuestionFeedback:
      type: object
      required: [correct, incorrect]
      properties:
        correct: { type: string, minLength: 1 }
        incorrect: { type: string, minLength: 1 }
    QuestionRubric:
      type: object
      required: [keyPoints, misconceptions]
      properties:
        keyPoints:
          type: array
          minItems: 1
          items: { type: string }
        misconceptions:
          type: array
          minItems: 1
          items: { type: string }
    Question:
      type: object
      required: [id, type, prompt, targetConcepts, feedback, rubrics]
      properties:
        id: { type: string, minLength: 1 }
        type: { $ref: '#/components/schemas/QuestionType' }
        prompt: { type: string, minLength: 1 }
        options:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/QuestionOption' }
          description: Required (and only validated) when `type` is `mcq`.
        correctOptionId:
          type: string
          description: Required (and only validated) when `type` is `mcq`.
        targetConcepts:
          type: array
          minItems: 1
          items: { type: string }
        feedback: { $ref: '#/components/schemas/QuestionFeedback' }
        rubrics: { $ref: '#/components/schemas/QuestionRubric' }
    QuestionSetType:
      type: string
      enum: [regular, live]
    PopulatedTopic:
      type: object
      description: >-
        Raw populated Mongoose topic document as serialized by
        `QuestionSetResponseDto` (`_id`, `__v` and the `category`
        ObjectId are NOT remapped).
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        icon: { type: string }
        tags:
          type: array
          items: { type: string }
        category: { $ref: '#/components/schemas/ObjectId' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        __v: { type: integer }
    QuestionSet:
      type: object
      required: [id, topic, setType, level, questions]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        topic:
          description: ObjectId hex string, or the populated topic.
          oneOf:
            - $ref: '#/components/schemas/ObjectId'
            - $ref: '#/components/schemas/PopulatedTopic'
        setType: { $ref: '#/components/schemas/QuestionSetType' }
        level: { type: integer, minimum: 0 }
        questions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Question' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateQuestionSetRequest:
      type: object
      required: [topic, setType, level, questions]
      additionalProperties: false
      properties:
        topic: { $ref: '#/components/schemas/ObjectId' }
        setType: { $ref: '#/components/schemas/QuestionSetType' }
        level:
          type: integer
          minimum: 0
          description: Required nonnegative whole-number level.
        questions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Question' }
    UpdateQuestionSetRequest:
      type: object
      additionalProperties: false
      properties:
        topic: { $ref: '#/components/schemas/ObjectId' }
        setType: { $ref: '#/components/schemas/QuestionSetType' }
        level:
          type: integer
          minimum: 0
          description: Nonnegative whole-number level when supplied.
        questions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Question' }
    BulkUpdateQuestionSetRequest:
      type: object
      required: [id]
      additionalProperties: false
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        topic: { $ref: '#/components/schemas/ObjectId' }
        setType: { $ref: '#/components/schemas/QuestionSetType' }
        level:
          type: integer
          minimum: 0
          description: Nonnegative whole-number level when supplied.
        questions:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Question' }
...